requirements.txt
web3
fastapi
cachetools
dataclasses
hexbytes
jwcrypto
spurplus
invoke
typing_extensions
uvicorn
aiofiles
jinja2
aiopyql
devtools
pygments
fastapi-simple-security
pillow
qrcode
pypng
httptools